// enums/employee/employee-hr-files.enum.ts
export enum EmployeeHRFiles {
    RESUME = 'RESUME',
    OFFER_LETTER = 'OFFER_LETTER',
    W9_FORM = 'W9_FORM',
    W4_FORM = 'W4_FORM',
    I9_FORM = 'I9_FORM',
    SECOND_ID = 'SECOND_ID',
    COMPANY_POLICIES_RECEIPT = 'COMPANY_POLICIES_RECEIPT',
    CONTROLLED_SUBSTANCE_POLICY_RECEIPT = 'CONTROLLED_SUBSTANCE_POLICY_RECEIPT',
    BANK_DEPOSIT_INFO = 'BANK_DEPOSIT_INFO',
    EMERGENCY_CONTACT_LIST = 'EMERGENCY_CONTACT_LIST',
    TRUCK_PROVIDED = 'TRUCK_PROVIDED',
}